forex_bot/forex_bot/requirements.txt
oandapyV20
ta
python-dotenv

